add article components & test authentication

+++ source/server/models/article/comment.js
@@ -2,9 +2,9 @@ const _ = require('loadsh');
 const Strings = require('../../utils/String');
 const query = require('../../models/query');
 
-const getComments = function(articleGuid) {
+const getComments = function (articleGuid) {
 	let sql = 'SELECT c.Guid, c.Content, c.CreationDate, c.Author FROM t_comment c '
-			+ 'WHERE c.ArticleGuid=? ORDER BY c.CreationDate';
+		+ 'WHERE c.ArticleGuid=? ORDER BY c.CreationDate';
 	return query.execute({
 		statement: sql,
 		params: [articleGuid]
@@ -20,21 +20,21 @@ const getComments = function(articleGuid) {
 const getComment = (commentGuid) => {
 	let sql = 'SELECT c.Guid, c.Content, c.CreationDate, c.Author, u.Username 		FROM t_comment c ' +
 		'LEFT JOIN (SELECT Username, Guid FROM t_user GROUP BY Username) AS u ON u.Guid = c.Author WHERE c.Guid=?';
-	
+
 	return query.execute({
-			statement: sql,
-			params: [commentGuid]
-		}).then(rs => {
-			if (_.isEmpty(rs)) {
-				return [];
-			} else {
-				return rs;
-			}
+		statement: sql,
+		params: [commentGuid]
+	}).then(rs => {
+		if (_.isEmpty(rs)) {
+			return [];
+		} else {
+			return rs;
+		}
 	});
-	
+
 }
 
-const addComment = function(comment) {
+const addComment = function (comment) {
 	var [columns, params] = createStatement(comment);
 	columns.push('CreationDate=?');
 	params.push(Strings.formatDate());
@@ -56,11 +56,11 @@ const addComment = function(comment) {
 };
 
 
-const deleteComment = function(commentGuid) {
-	let sql = 'DELETE FROM t_comment WHERE Guid=?';
+const deleteComment = function (commentGuid, userGuid) {
+	let sql = 'DELETE FROM t_comment WHERE Guid=? AND Author=?';
 	return query.execute({
 		statement: sql,
-		params: [commentGuid]
+		params: [commentGuid, userGuid]
 	}).then(rs => {
 		if (rs.affectedRows === 1) {
 			return rs.affectedRows;
@@ -70,6 +70,15 @@ const deleteComment = function(commentGuid) {
 	});
 };
 
+const count = (articleGuid) => {
+	let sql = "SELECT COUNT(id) AS Count WHERE ArticleGuid=?";
+	return query.execute({
+		statement: sql,
+		params: [articleGuid]
+	}).then(rs => {
+		return rs[0].Count;
+	});
+}
 function createStatement(comment) {
 	let columns = [];
 	let params = [];
@@ -89,4 +98,5 @@ module.exports.commentModel = {
 	getComment,
 	addComment,
 	deleteComment,
+	count,
 }
